Rename `DefaultLogger` to `ShipBookLogger`
